+++ R/MeasureSurvMAE.R
@@ -21,7 +21,7 @@ MeasureSurvMAE = R6::R6Class("MeasureSurvMAE",
     #' @description Creates a new instance of this [R6][R6::R6Class] class.
     initialize = function(se = FALSE) {
       super$initialize(
-        id = "surv.mae",
+        id = ifelse(se, "surv.mae_se", "surv.mae"),
         range = c(0, Inf),
         minimize = TRUE,
         predict_type = "response",