+++ R/MeasureSurvNagelkR2.R
@@ -9,7 +9,7 @@
 #' @template measure_survAUC
 #'
 #' @references
-#' \cite{mlr3proba}{nagelkerke_1991}
+#' `r tools::toRd(bibentries["nagelkerke_1991"])`
 #'
 #' @family R2 survival measures
 #' @family lp survival measures