+++ NEWS.md
@@ -1,8 +1,12 @@
 # mlr3proba 0.2.2.9000
 
+* All learners that previously lived in the mlr3learners organisation are now in the [mlr3extralearners](https://github.com/mlr-org/mlr3extralearners) repository.
 * Fixed bottleneck in `MeasureSurvLogloss`
 * Bugfix in `MeasureSurvCalibrationAlpha`
 * Removed `pracma` from `Suggests`
+* Patch for valgrind
+* `TaskDens` now inherits from `TaskUnsupervised`
+* Added calibration plot for comparing average predicted survival distribution to Kaplan-Meier
 
 # mlr3proba 0.2.2