+++ NEWS.md
@@ -3,6 +3,7 @@
 * All learners that previously lived in the mlr3learners organisation are now in the [mlr3extralearners](https://github.com/mlr-org/mlr3extralearners) repository.
 * Fixed bottleneck in `MeasureSurvLogloss`
 * Bugfix in `MeasureSurvCalibrationAlpha`
+* Removed `pracma` from `Suggests`
 * Patch for valgrind
 * `TaskDens` now inherits from `TaskUnsupervised`
 * Added calibration plot for comparing average predicted survival distribution to Kaplan-Meier